Add categories and rename first post
The name of the first post has bothered me for a bit.

Also I recently (re)discovered the categories are empty which makes the
Categories page empty.  I considered removing it but eh...  this is
easier for now and most of my stuff fits in a few categories anyway.

Maybe I'll add arts and linkdump.